move discoverSchemaForSourceFromSourceId to sourceHandler
